Zivid: Add function to reinterpret Buffers
Buffers can now be reinterpreted into tother types and dimensionality
using Buffer::reinterpret. This function checks whether the conversion
is safe before performing it and issues either a compile time or runtime
error if it's not.

The use case is when we want to cache buffers without knwing their type
upfront.

MISC